requirements.txt
fastapi
uvicorn[standard]
qrcode[pil]
python-multipart
Pillow
